refactor(main): move pipeline, ipcs out of factories dir
