+++ src/main/factories/index.ts
@@ -1,4 +1,2 @@
-export * from './ipcs/register-window-creation'
-export * from './ipcs/pipeline'
 export * from './windows/create'
 export * from './app'
