--- src/main/factories/ipcs/index.ts
@@ -1,3 +1,0 @@
-export * from './register-window-creation'
-export * from '../windows/create'
-export * from '../app'
