+++ src/main/ipcs/browser.ts
@@ -1,7 +1,7 @@
 import { ipcMain, shell } from 'electron'
 
 // helper functions
-import { IPC_EVENT_openInBrowser } from '../shared/main-renderer-events'
+import { IPC_EVENT_openInBrowser } from '../../shared/main-renderer-events'
 
 function setupOpenInBrowserEvents() {
     // payload should be "/Path/to/folder" not file://..
