+++ src/main/ipcs/clipboard.ts
