+++ src/main/ipcs/index.ts
@@ -1,0 +1,3 @@
+export * from './register-window-creation'
+export * from '../factories/windows/create'
+export * from '../factories/app'
