+++ src/main/ipcs/one-time-fetch.ts
@@ -1,7 +1,7 @@
 // fetch a resource and return the result to the renderer
 // this is meant as a one-off utility for special cases
 import { ipcMain } from 'electron'
-import { IPC_EVENT_oneTimeFetch } from '../shared/main-renderer-events'
+import { IPC_EVENT_oneTimeFetch } from '../../shared/main-renderer-events'
 import fetch from 'node-fetch'
 
 async function oneTimeFetch(url) {
