+++ src/main/ipcs/register-window-creation.ts
